# 8344766: AES/CTR slow at big payloads