This is a follow up to https://github.com/openjdk/jdk/pull/22086 for AES/CTR

Before:
```
Benchmark                (algorithm)  (dataSize)  (keyLength)  (provider)   Mode  Cnt    Score    Error  Units
AESBench.decrypt   AES/CTR/NoPadding    30000000          128      SunJCE  thrpt    3   16.491 ±  0.356  ops/s
AESBench.decrypt2  AES/CTR/NoPadding    30000000          128      SunJCE  thrpt    3   16.899 ±  0.013  ops/s
AESBench.encrypt   AES/CTR/NoPadding    30000000          128      SunJCE  thrpt    3   16.477 ±  1.006  ops/s
AESBench.encrypt2  AES/CTR/NoPadding    30000000          128      SunJCE  thrpt    3   16.921 ±  0.038  ops/s
```
After:
```
Benchmark                (algorithm)  (dataSize)  (keyLength)  (provider)   Mode  Cnt    Score   Error  Units
AESBench.decrypt   AES/CTR/NoPadding    30000000          128      SunJCE  thrpt    3  218.910 ± 1.991  ops/s
AESBench.decrypt2  AES/CTR/NoPadding    30000000          128      SunJCE  thrpt    3  426.414 ± 2.988  ops/s
AESBench.encrypt   AES/CTR/NoPadding    30000000          128      SunJCE  thrpt    3  218.882 ± 2.446  ops/s
AESBench.encrypt2  AES/CTR/NoPadding    30000000          128      SunJCE  thrpt    3  425.402 ± 4.205  ops/s
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8344766](https://bugs.openjdk.org/browse/JDK-8344766): AES/CTR slow at big payloads (**Enhancement** - P4)


### Reviewers
 * [Anthony Scarpino](https://openjdk.org/census#ascarpino) (@ascarpino - **Reviewer**)
 * [Jatin Bhateja](https://openjdk.org/census#jbhateja) (@jatin-bhateja - **Reviewer**)

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/22300/head:pull/22300` \
`$ git checkout pull/22300`

Update a local copy of the PR: \
`$ git checkout pull/22300` \
`$ git pull https://git.openjdk.org/jdk.git pull/22300/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 22300`

View PR using the GUI difftool: \
`$ git pr show -t 22300`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/22300.diff">https://git.openjdk.org/jdk/pull/22300.diff</a>

</details>
<details><summary>Using Webrev</summary>

[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/22300#issuecomment-2492011333)
</details>